+++ src/components/Header/Header.test.js
@@ -9,9 +9,7 @@ jest.mock('axios');
 
 jest.mock('react-router-dom', () => ({
   ...jest.requireActual('react-router-dom'),
-  useHistory: () => ({
-    push: jest.fn()
-  })
+  useNavigate: () => jest.fn()
 }));
 
 jest.mock('../Editor/EditorSlice', () => ({