+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Upgrade `react-router` to `v6`
+
 ### Fixed
 
 - Show `p5` error messages in the user interface (#346)